requirements.txt
# Core Django packages
Django==5.2.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
mysqlclient==2.2.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==3.0.1
social-auth-app-django==5.4.0

# File handling & Media
Pillow==11.2.1
django-storages==1.14.2
boto3==1.38.27
cloudinary==1.44.0
django-cloudinary-storage==0.3.0

# Email
django-anymail==13.0

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# API Documentation
drf-spectacular==0.28.0

# Testing
pytest==8.4.0
pytest-django==4.7.0
factory-boy==3.3.0

# Utilities
django-extensions==4.1
django-filter==25.1
django-crispy-forms==2.1

# AI Integration
google-generativeai==0.8.5

# Production
gunicorn==23.0.0
whitenoise==6.6.0

# Development
django-debug-toolbar==5.2.0

# Caching
redis==6.2.0
django-redis==5.4.0

# Data validation
django-phonenumber-field==8.1.0
phonenumbers==9.0.6

# Timezone support
pytz==2025.2

# Password hashing
argon2-cffi==23.1.0

# File validation
python-magic==0.4.27

# Rate limiting
django-ratelimit==4.1.0

# Monitoring
sentry-sdk==2.29.1
